Use int32 instead of int for CRD integer fields
Kubernetes API conventions require int32 or int64 for integer fields.
Go's int is platform-dependent and generates format: int64 in the
OpenAPI schema. All 8 affected fields represent small values (counts,
port numbers) that fit in int32 and should match K8s conventions.

Changed fields:
- VirtualMCPServerStatus.BackendCount
- MCPGroupStatus.ServerCount, RemoteProxyCount
- MCPRegistryDatabaseConfig.Port, MaxOpenConns, MaxIdleConns
- SyncStatus.AttemptCount, ServerCount

Fixes #4544

+++ cmd/thv-operator/test-integration/mcp-group/mcpgroup_controller_integration_test.go
@@ -148,7 +148,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 
 		It("Should find existing MCPServers and update status", func() {
 			// Check that the group found both servers
-			Eventually(func() int {
+			Eventually(func() int32 {
 				updatedGroup := &mcpv1alpha1.MCPGroup{}
 				if err := k8sClient.Get(ctx, types.NamespacedName{
 					Name:      mcpGroupName,
@@ -157,7 +157,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 					return -1
 				}
 				return updatedGroup.Status.ServerCount
-			}, timeout, interval).Should(Equal(2))
+			}, timeout, interval).Should(Equal(int32(2)))
 
 			// The group should be Ready after successful reconciliation
 			Eventually(func() mcpv1alpha1.MCPGroupPhase {
@@ -421,7 +421,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 
 			// The MCPGroup should be Ready with 2 servers
 			Expect(updatedGroup.Status.Phase).To(Equal(mcpv1alpha1.MCPGroupPhaseReady))
-			Expect(updatedGroup.Status.ServerCount).To(Equal(2))
+			Expect(updatedGroup.Status.ServerCount).To(Equal(int32(2)))
 
 			// Trigger a reconciliation by updating the MCPGroup spec
 			Eventually(func() error {
@@ -525,7 +525,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 			Expect(k8sClient.Create(ctx, mcpGroup)).Should(Succeed())
 
 			// Wait for initial reconciliation - the group should find the servers
-			Eventually(func() int {
+			Eventually(func() int32 {
 				updatedGroup := &mcpv1alpha1.MCPGroup{}
 				if err := k8sClient.Get(ctx, types.NamespacedName{
 					Name:      mcpGroupName,
@@ -534,7 +534,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 					return -1
 				}
 				return updatedGroup.Status.ServerCount
-			}, timeout, interval).Should(Equal(2))
+			}, timeout, interval).Should(Equal(int32(2)))
 		})
 
 		AfterAll(func() {
@@ -593,7 +593,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 				Name:      mcpGroupName,
 				Namespace: namespace,
 			}, updatedGroup)).Should(Succeed())
-			Expect(updatedGroup.Status.ServerCount).To(Equal(2))
+			Expect(updatedGroup.Status.ServerCount).To(Equal(int32(2)))
 		})
 	})
 
@@ -725,7 +725,7 @@ var _ = Describe("MCPGroup Controller Integration Tests", func() {
 				Namespace: namespaceA,
 			}, updatedGroup)).Should(Succeed())
 
-			Expect(updatedGroup.Status.ServerCount).To(Equal(1))
+			Expect(updatedGroup.Status.ServerCount).To(Equal(int32(1)))
 			Expect(updatedGroup.Status.Servers).To(ContainElement("server-a"))
 			Expect(updatedGroup.Status.Servers).NotTo(ContainElement("server-b"))
 		})
